Add HomeScreen with tabs and integrate NewJobScreen
